Replace ayeinput by ayemouse
Given [1], there is no longer a need to set the keyboard layout using
XKB options and xmodmap. Also, given the switch from xf86-input-evdev to
xf86-input-libinput, scrolling with the middle mouse button seems to
have become standard behaviour.

Consequently, all we need to do is simply check whether we're using a
ThinkPad keyboard, and swap the left and right mouse buttons if so.

[1] https://github.com/ayekat/PKGBUILDs/commit/695dcf525fcd37ad0a9f0dd5ef02e86225df5b2e