Add pooch
pooch is a script that watches a Maildir and does 2 things when new mail
arrives in the inbox:

* Apply filter rules from $XDG_CONFIG_HOME/pooch to move the mails to
  their respective folders within the account's Maildir.
* Send a notification via libnotify.

It is a poor-man's procmail for people who filter mails on the
client-side (synchronised with mbsync).

There are still some todos and unhandled cases that result in mails not
being properly matched, but it shouldn't break things, and I consider it
to be stable enough at least for my own private use (currently being
tested).