Remove pacman helper scripts, make vidcat bash
There is no point in keeping pacman-related scripts here for two
reasons:

* The utils repository is rather general-purpose, not
  pacman/Arch-specific.
* The argument of "we can have them at our fingertips without being
  required to have root rights" doesn't make any sense, because those
  scripts are only meant to be run by a system administrator anyway.

Putting them in a separate repo and packaging them as an Arch Linux
package therefore makes more sense. We also get rid of a few
dependencies (which is great, because we can't handle them anyway on
Debian systems).